requirements.txt
torch
torchvision
torchaudio
pandas
numpy
Pillow
tqdm
